+++ example/src/components/index.ts
@@ -1,0 +1,8 @@
+import "./index.css";
+import FadeInComponent from '@components/FadeInComponent/FadeInComponent';
+import Page from '@components/Page/Page';
+
+export {
+  FadeInComponent,
+  Page,
+};