+++ example/src/providers/GlobalProvider.tsx
@@ -1,0 +1,14 @@
+// Global provider
+// Path: src/providers/GlobalProvider.ts
+
+import { createContext } from "react";
+
+interface GlobalContextProps { }
+
+export const GlobalContext = createContext<GlobalContextProps>({});
+
+export default function GlobalProvider({ children }: {
+  children: React.ReactNode;
+}) {
+  return <GlobalContext.Provider value={{}}>{children}</GlobalContext.Provider>;
+}