+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ### Version 0.4.0
     - Major refactoring to clean up the code
+    - Added example project
     - Fix ensureVisibility ...
 
 ### Version 0.3.x